// src/types/patients.ts
export interface Patient {
  id: number;
  name: string;
  gender: string;
  date_of_birth: string;
  age: number;
  summary: string;
  email: string;
  phone: string;
  address: string;
  insurance_member_id: string;
  policy_holder: string;
  additional_info: string;
  height: number;
  weight: number;
  blood_group: string;
  created_at: string;
  updated_at: string;
}